+++ keyboards/ferris/0_1/0_1.c
@@ -14,4 +14,4 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "0.1.h"
+#include "0_1.h"
