+++ keyboards/ferris/readme.md
@@ -13,7 +13,7 @@ A split 34 keys column staggered keyboard named and decorated after the rustlang
 
 Make examples for this keyboard (after setting up your build environment):
 
-    make ferris/0.1:default
+    make ferris/0_1:default
     make ferris/sweep:default:avrdude-split-right
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
